Remove HLS256 and HSV256 support (see issue #1)
Because Hue is given in degrees (of 360) HLS and HSV 256 support
does not make sense. Only floats are available for returning, in the
format (hue: percent of 360, lightness: percent of 100, saturation:
percent of 100).

Nudge numbers in unit test to convert properly.

Fix bug in HSV.setter.

Fix typo in chroma.Color.